# [SYSTEMML-445] Improved performance of GPU right indexing