--- a/src/main/java/org/apache/sysml/api/DMLScript.java
+++ b/src/main/java/org/apache/sysml/api/DMLScript.java
@@ -173,7 +173,10 @@ public class DMLScript
 	
 	public static boolean           USE_ACCELERATOR     = DMLOptions.defaultOptions.gpu;
 	public static boolean           FORCE_ACCELERATOR   = DMLOptions.defaultOptions.forceGPU;
-	public static boolean           SYNCHRONIZE_GPU  	= true;   									// whether to synchronize GPU after every instruction
+	// whether to synchronize GPU after every instruction
+	public static boolean           SYNCHRONIZE_GPU  	= true;
+	// whether to perform eager CUDA free on rmvar
+	public static boolean           EAGER_CUDA_FREE  	= false;
 
 
 	public static boolean _suppressPrint2Stdout = false;  // flag that indicates whether or not to suppress any prints to stdout