--- a/src/main/java/org/apache/sysml/conf/DMLConfig.java
+++ b/src/main/java/org/apache/sysml/conf/DMLConfig.java
@@ -83,7 +83,8 @@ public class DMLConfig
 	public static final String EXTRA_GPU_STATS      = "systemml.stats.extraGPU"; //boolean
 	public static final String EXTRA_DNN_STATS      = "systemml.stats.extraDNN"; //boolean
 	public static final String AVAILABLE_GPUS       = "systemml.gpu.availableGPUs"; // String to specify which GPUs to use (a range, all GPUs, comma separated list or a specific GPU)
-	public static final String SYNCHRONIZE_GPU       = "systemml.gpu.sync.postProcess"; // boolean: whether to synchronize GPUs after every instruction 
+	public static final String SYNCHRONIZE_GPU      = "systemml.gpu.sync.postProcess"; // boolean: whether to synchronize GPUs after every instruction 
+	public static final String EAGER_CUDA_FREE		= "systemml.gpu.eager.cudaFree"; // boolean: whether to perform eager CUDA free on rmvar
 	
 	// Fraction of available memory to use. The available memory is computer when the GPUContext is created
 	// to handle the tradeoff on calling cudaMemGetInfo too often.
@@ -132,7 +133,8 @@ public class DMLConfig
 		_defaultVals.put(EXTRA_DNN_STATS,        "false" );
 		_defaultVals.put(GPU_MEMORY_UTILIZATION_FACTOR,      "0.9" );
 		_defaultVals.put(AVAILABLE_GPUS,         "-1");
-		_defaultVals.put(SYNCHRONIZE_GPU,        "false" );
+		_defaultVals.put(SYNCHRONIZE_GPU,        "true" );
+		_defaultVals.put(EAGER_CUDA_FREE,        "false" );
 	}
 	
 	public DMLConfig()
@@ -415,7 +417,7 @@ public class DMLConfig
 				COMPRESSED_LINALG, 
 				CODEGEN, CODEGEN_COMPILER, CODEGEN_PLANCACHE, CODEGEN_LITERALS,
 				EXTRA_GPU_STATS, EXTRA_DNN_STATS, EXTRA_FINEGRAINED_STATS, STATS_MAX_WRAP_LEN,
-				AVAILABLE_GPUS, SYNCHRONIZE_GPU
+				AVAILABLE_GPUS, SYNCHRONIZE_GPU, EAGER_CUDA_FREE
 		}; 
 		
 		StringBuilder sb = new StringBuilder();