--- a/src/main/java/org/apache/sysml/runtime/instructions/gpu/context/CSRPointer.java
+++ b/src/main/java/org/apache/sysml/runtime/instructions/gpu/context/CSRPointer.java
@@ -521,7 +521,7 @@ public class CSRPointer {
 	 * @throws DMLRuntimeException ?
 	 */
 	public void deallocate() throws DMLRuntimeException {
-		deallocate(false);
+		deallocate(DMLScript.EAGER_CUDA_FREE);
 	}
 
 	/**