--- a/src/main/java/org/apache/sysml/runtime/instructions/gpu/context/GPUContext.java
+++ b/src/main/java/org/apache/sysml/runtime/instructions/gpu/context/GPUContext.java
@@ -359,7 +359,7 @@ public class GPUContext {
 	 * @param toFree {@link Pointer} instance to be freed
 	 */
 	public void cudaFreeHelper(final Pointer toFree) {
-		cudaFreeHelper(null, toFree, false);
+		cudaFreeHelper(null, toFree, DMLScript.EAGER_CUDA_FREE);
 	}
 
 	/**
@@ -379,7 +379,7 @@ public class GPUContext {
 	 * @param toFree          {@link Pointer} instance to be freed
 	 */
 	public void cudaFreeHelper(String instructionName, final Pointer toFree) {
-		cudaFreeHelper(instructionName, toFree, false);
+		cudaFreeHelper(instructionName, toFree, DMLScript.EAGER_CUDA_FREE);
 	}
 
 	/**