--- a/src/main/java/org/apache/sysml/runtime/matrix/data/LibMatrixCUDA.java
+++ b/src/main/java/org/apache/sysml/runtime/matrix/data/LibMatrixCUDA.java
@@ -2249,14 +2249,13 @@ public class LibMatrixCUDA {
 					+ (cl + 1) + ":" + (cu + 1) + "] " + "must be within matrix dimensions [" + in1.getNumRows() + ","
 					+ in1.getNumColumns() + "]");
 		}
-
-		
+		int len1 = toInt(in1.getNumColumns());
 		if(isInSparseFormat(gCtx, in1)) {
 			// Input in1 is in sparse format and output is in dense format
 			MatrixObject out = getDenseMatrixOutputForGPUInstruction(ec, instName, outputName, ru - rl + 1, cu - cl + 1);
 			CSRPointer inPointer = getSparsePointer(gCtx, in1, instName);
 			Pointer outPointer = getDensePointer(gCtx, out, instName);
-			sliceSparseDense(gCtx, instName, inPointer, outPointer, rl, ru, cl, cu);
+			sliceSparseDense(gCtx, instName, inPointer, outPointer, rl, ru, cl, cu, len1);
 		}
 		else {
 			// Input in1 is in dense format (see inPointer)
@@ -2264,9 +2263,7 @@ public class LibMatrixCUDA {
 
 			Pointer inPointer = getDensePointer(gCtx, in1, instName);
 			Pointer outPointer = getDensePointer(gCtx, out, instName);
-			int len1 = toInt(in1.getNumColumns());
-			int len2 = toInt(ec.getMatrixObject(outputName).getNumColumns());
-			sliceDenseDense(gCtx, instName, inPointer, outPointer, rl, ru, cl, cu, len1, len2);
+			sliceDenseDense(gCtx, instName, inPointer, outPointer, rl, ru, cl, cu, len1);
 		}
 	}
 	
@@ -2282,19 +2279,19 @@ public class LibMatrixCUDA {
 	 * @param cl column lower
 	 * @param cu column upper
 	 * @param inClen input number of columns
-	 * @param retClen output number of columns
 	 * @throws DMLRuntimeException if error occurs
 	 */
 	protected static void sliceDenseDense(GPUContext gCtx, String instName, Pointer inPointer, Pointer outPointer, 
-			int rl, int ru, int cl, int cu, int inClen, int retClen) throws DMLRuntimeException {
+			int rl, int ru, int cl, int cu, int inClen) throws DMLRuntimeException {
 		long t0 = GPUStatistics.DISPLAY_STATISTICS ? System.nanoTime() : 0;
+		long retClen = cu - cl + 1;
 		if (inClen == retClen) {
 			cudaMemcpy(outPointer, inPointer.withByteOffset(rl * inClen * Sizeof.DOUBLE), (ru - rl + 1) * inClen
 					* Sizeof.DOUBLE, cudaMemcpyDeviceToDevice);
 		} else {
-			int size = ru - rl + 1;
-			getCudaKernels(gCtx).launchKernel("slice_dense_dense", ExecutionConfig.getConfigForSimpleVectorOperations(size),
-					inPointer, outPointer, rl, ru, cl, cu, inClen, retClen);
+			long retRlen = ru - rl + 1;
+			getCudaKernels(gCtx).launchKernel("slice_dense_dense", ExecutionConfig.getConfigForSimpleVectorOperations(toInt(retRlen*retClen)),
+					inPointer, outPointer, rl, ru, cl, cu, inClen,  retRlen, retClen);
 		}
 		if (GPUStatistics.DISPLAY_STATISTICS) GPUStatistics.maintainCPMiscTimes(instName, GPUInstruction.MISC_TIMER_RIX_DENSE_OP, System.nanoTime() - t0);
 	}
@@ -2310,18 +2307,37 @@ public class LibMatrixCUDA {
 	 * @param ru row upper
 	 * @param cl column lower
 	 * @param cu column upper
+	 * @param inClen number of columns of input matrix
 	 * @throws DMLRuntimeException if error
 	 */
-	protected static void sliceSparseDense(GPUContext gCtx, String instName, CSRPointer inPointer, Pointer outPointer, int rl, int ru, int cl, int cu) throws DMLRuntimeException {
-		int size = ru - rl + 1;
+	protected static void sliceSparseDense(GPUContext gCtx, String instName, CSRPointer inPointer, Pointer outPointer, 
+			int rl, int ru, int cl, int cu, int inClen) throws DMLRuntimeException {
+		int retRlen = ru - rl + 1;
 		long t0 = GPUStatistics.DISPLAY_STATISTICS ? System.nanoTime() : 0;
 		int retClen = cu - cl + 1;
+		
+		int size = -1; String kernel = null; String timer = null;
+		if(retClen > 2*retRlen) {
+			int[] rlPtr = { -1 }; int[] ruPtr = { -1 };
+			cudaMemcpy(Pointer.to(rlPtr), inPointer.rowPtr.withByteOffset(rl*Sizeof.INT), Sizeof.INT, cudaMemcpyDeviceToHost);
+			cudaMemcpy(Pointer.to(ruPtr), inPointer.rowPtr.withByteOffset((ru+1)*Sizeof.INT), Sizeof.INT, cudaMemcpyDeviceToHost);
+			size = ruPtr[0] - rlPtr[1];
+			timer = GPUInstruction.MISC_TIMER_RIX_SPARSE_DENSE_OP_NNZ;
+			kernel = "slice_sparse_dense_nnz";
+		}
+		else {
+			size = retRlen;
+			timer = GPUInstruction.MISC_TIMER_RIX_SPARSE_DENSE_OP_ROWWISE;
+			kernel = "slice_sparse_dense_row";
+		}
+		
+		
 		// Performs a slice operation where the input matrix is sparse and the output matrix is dense.
 		// This function avoids unnecessary sparse to dense conversion of the input matrix.
 		// We can generalize this later to output sparse matrix.
-		getCudaKernels(gCtx).launchKernel("slice_sparse_dense", ExecutionConfig.getConfigForSimpleVectorOperations(size),
+		getCudaKernels(gCtx).launchKernel(kernel, ExecutionConfig.getConfigForSimpleVectorOperations(size),
 				inPointer.val, inPointer.rowPtr, inPointer.colInd, outPointer, rl, ru, cl, cu, retClen);
-		if (GPUStatistics.DISPLAY_STATISTICS) GPUStatistics.maintainCPMiscTimes(instName, GPUInstruction.MISC_TIMER_RIX_SPARSE_DENSE_OP, System.nanoTime() - t0);
+		if (GPUStatistics.DISPLAY_STATISTICS) GPUStatistics.maintainCPMiscTimes(instName, timer, System.nanoTime() - t0);
 	}
 
 	public static void cbind(ExecutionContext ec, GPUContext gCtx, String instName, MatrixObject in1, MatrixObject in2, String outputName) throws DMLRuntimeException {