- Improved performance of dense right-indexing by increasing parallelization to elements rather than row.
- Added slice_sparse_dense_nnz to handle indexing of sparse wide matrices.
- Added EAGER_CUDA_FREE flag to allow users to enable/disable cudaFree optimization.
- Introduced IGNORE_CLEAR_MEMORY_BUG flag in GPUTests to run tests.

@nakul02 can you please review this PR ?

As shown in the below numbers, this PR shows performance improvement in almost all scenarios and especially for large cases. 

 
## Setup 1: right indexing microbenchmarks

```
rl = ceil($dim1*0.1); ru = ceil($dim1*0.3); cl = ceil($dim2*0.1); cu = ceil($dim2*0.3)
X = rand(rows=$dim1, cols=$dim2, sparsity=$sp)
s = 0
for(i in 1:100) {
        s = s + sum(X[rl:ru,cl:cu])
}
print(s)
```

###  How to read the below numbers?
The first line corresponds to the heavy hitter from this PR and the second line corresponds to the heavy hitter from the current master. 

`drix`: dense right indexing kernel, `sdrixn`: sparse dense right indexing kernel with nnz parallelization, 
 `sdrixr`: sparse dense right indexing kernel with row-wise parallelization and `sdrix`: previous sparse dense right indexing kernel.



Performance results:
```bash
[100, 100] sparsity=0.01
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.030    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixr[0.011s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.030    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.011s,100]

[100, 100] sparsity=0.1
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.030    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrixr[0.012s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.034    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrix[0.014s,100]

[100, 100] sparsity=0.2
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.033    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixr[0.014s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.034    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.015s,100]

[100, 100] sparsity=0.5
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.029    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.013s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.029    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.013s,100]

[100, 100] sparsity=0.9
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.030    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.014s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.031    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.014s,100]

[100, 1000] sparsity=0.01
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.039    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrixn[0.020s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.033    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrix[0.013s,100]

[100, 1000] sparsity=0.1
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.038    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrixn[0.019s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.046    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrix[0.028s,100]

[100, 1000] sparsity=0.2
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.040    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrixn[0.020s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.065    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrix[0.043s,100]

[100, 1000] sparsity=0.5
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.031    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.015s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.054    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.038s,100]

[100, 1000] sparsity=0.9
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.033    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.004s,100], drix[0.015s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.055    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.039s,100]

[100, 5000] sparsity=0.01
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.042    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixn[0.023s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.039    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.019s,100]

[100, 5000] sparsity=0.1
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.040    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixn[0.020s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.112    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.091s,100]

[100, 5000] sparsity=0.2
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.041    100                                    aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrixn[0.020s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.181    100                                     aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrix[0.160s,100]

[100, 5000] sparsity=0.5
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.033    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.014s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.171    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.151s,100]

[100, 5000] sparsity=0.9
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.030    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.012s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.167    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.150s,100]

[1000, 100] sparsity=0.01
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.030    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrixr[0.011s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.032    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrix[0.011s,100]

[1000, 100] sparsity=0.1
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.033    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrixr[0.013s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.031    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrix[0.012s,100]

[1000, 100] sparsity=0.2
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.041    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixr[0.016s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.034    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.003s,100], sdrix[0.015s,100]

[1000, 100] sparsity=0.5
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.031    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.013s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.032    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.015s,100]

[1000, 100] sparsity=0.9
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.030    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.013s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.031    100                                      aqrd[0.000s,1], H2D[0.000s,1], ad[0.003s,100], drix[0.015s,100]

[1000, 1000] sparsity=0.01
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.033    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixr[0.013s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.032    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.012s,100]

[1000, 1000] sparsity=0.1
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.054    100                                    aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrixr[0.032s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.054    100                                     aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrix[0.033s,100]

[1000, 1000] sparsity=0.2
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.076    100                                    aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrixr[0.053s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.081    100                                     aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrix[0.053s,100]

[1000, 1000] sparsity=0.5
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.032    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.013s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.069    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.051s,100]

[1000, 1000] sparsity=0.9
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.031    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.013s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.070    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.052s,100]

[1000, 5000] sparsity=0.01
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.052    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixn[0.029s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.045    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.024s,100]

[1000, 5000] sparsity=0.1
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.079    100                                    aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrixn[0.042s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.143    100                                     aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrix[0.120s,100]

[1000, 5000] sparsity=0.2
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.088    100                                    aqrs[0.000s,1], H2D[0.002s,1], ad[0.004s,100], sdrixn[0.060s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.246    100                                     aqrs[0.000s,1], H2D[0.002s,1], ad[0.004s,100], sdrix[0.217s,100]

[1000, 5000] sparsity=0.5
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.040    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.008s,1], drix[0.014s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.244    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.006s,1], drix[0.221s,100]

[1000, 5000] sparsity=0.9
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.039    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.006s,1], drix[0.015s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.238    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.006s,1], drix[0.213s,100]

[5000, 100] sparsity=0.01
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.032    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixr[0.011s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.032    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.011s,100]

[5000, 100] sparsity=0.1
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.036    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixr[0.015s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.036    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.014s,100]

[5000, 100] sparsity=0.2
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.047    100                                    aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrixr[0.020s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.043    100                                     aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrix[0.020s,100]

[5000, 100] sparsity=0.5
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.029    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.012s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.041    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.024s,100]

[5000, 100] sparsity=0.9
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.034    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.015s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.040    100                                      aqrd[0.000s,1], H2D[0.001s,1], ad[0.004s,100], drix[0.023s,100]

[5000, 1000] sparsity=0.01
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.037    100                                    aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrixr[0.014s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.036    100                                     aqrs[0.000s,1], H2D[0.000s,1], ad[0.004s,100], sdrix[0.014s,100]

[5000, 1000] sparsity=0.1
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.093    100                                    aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrixr[0.065s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.092    100                                     aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrix[0.064s,100]

[5000, 1000] sparsity=0.2
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.149    100                                    aqrs[0.000s,1], H2D[0.002s,1], ad[0.004s,100], sdrixr[0.117s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.150    100                                     aqrs[0.000s,1], H2D[0.002s,1], ad[0.004s,100], sdrix[0.118s,100]

[5000, 1000] sparsity=0.5
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.037    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.006s,1], drix[0.014s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.159    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.008s,1], drix[0.133s,100]

[5000, 1000] sparsity=0.9
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.038    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.006s,1], drix[0.014s,100]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.160    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.009s,1], drix[0.133s,100]

[5000, 5000] sparsity=0.01
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.063    100                                    aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrixr[0.038s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.063    100                                     aqrs[0.000s,1], H2D[0.001s,1], ad[0.004s,100], sdrix[0.039s,100]

[5000, 5000] sparsity=0.1
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.300    100                                    aqrs[0.000s,1], ad[0.004s,100], H2D[0.005s,1], sdrixr[0.260s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.307    100                                     aqrs[0.000s,1], ad[0.004s,100], H2D[0.005s,1], sdrix[0.269s,100]

[5000, 5000] sparsity=0.2
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.556    100                                    aqrs[0.000s,1], ad[0.004s,100], H2D[0.011s,1], sdrixr[0.490s,100]
  1  gpu_rightIndex [temp.dml 8:13-8:26]    0.539    100                                     aqrs[0.000s,1], ad[0.004s,100], H2D[0.008s,1], sdrix[0.478s,100]

[5000, 5000] sparsity=0.5
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.072    100                                      aqrd[0.000s,1], ad[0.004s,100], drix[0.024s,100], H2D[0.029s,1]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.597    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.038s,1], drix[0.539s,100]

[5000, 5000] sparsity=0.9
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.080    100                                      aqrd[0.000s,1], ad[0.005s,100], drix[0.023s,100], H2D[0.037s,1]
  2  gpu_rightIndex [temp.dml 8:13-8:26]    0.589    100                                      aqrd[0.000s,1], ad[0.004s,100], H2D[0.034s,1], drix[0.537s,100]
```

## Setup 2: 1-layer sparse CNN

This PR:
```
ITER=9 BEG=401 END=450 OBJ=1.9760828227695086
SystemML Statistics:
Total elapsed time:             32.148 sec.
Total compilation time:         16.677 sec.
Total execution time:           15.472 sec.
Number of compiled Spark inst:  54.
Number of executed Spark inst:  1.
CUDA/CuLibraries init time:     14.141/0.692 sec.
Number of executed GPU inst:    985.
GPU mem tx time  (alloc/dealloc/set0/toDev/fromDev):    0.161/0.154/0.014/0.191/0.044 sec.
GPU mem tx count (alloc/dealloc/set0/toDev/fromDev/evict):      349/333/1444/46/235/84/0.
GPU conversion time  (sparseConv/sp2dense/dense2sp):    0.075/0.030/0.000 sec.
GPU conversion count (sparseConv/sp2dense/dense2sp):    46/45/0.
Cache hits (Mem, WB, FS, HDFS): 195/0/0/2.
Cache writes (WB, FS, HDFS):    166/0/12.
Cache times (ACQr/m, RLS, EXP): 8.437/0.003/0.032/0.163 sec.
HOP DAGs recompiled (PRED, SB): 0/36.
HOP DAGs recompile time:        0.128 sec.
Spark ctx create time (lazy):   1.042 sec.
Spark trans counts (par,bc,col):0/0/0.
Spark trans times (par,bc,col): 0.000/0.000/0.000 secs.
Total JIT compile time:         10.094 sec.
Total JVM GC count:             21.
Total JVM GC time:              3.784 sec.
Heavy hitter instructions:
  #  Instruction                                         Time(s)  Count  Misc Timers
  1  rightIndex [cnn.dml 61:6-61:6]                        8.604      9                                      aqrs[8.140s,9], aqms[0.000s,9], rlswr[0.000s,0], rlsi[0.000s,9]
  2  gpu_conv2d_backward_filter [125:2-126:81]             2.832     36 nnc[0.001s,1800], ad[0.003s,36], nni[0.013s,1800], drix[0.031s,1800], nncbf[0.057s,1800], az[0.136s,
                                                                                                                                                  3600], sdrixn[0.235s,1800]
  3  forward [cnn.dml 61:6-61:6]                           2.556      9
  4  gpu_conv2d_bias_add [78:2-78:23]                      2.202     36 aqrs[0.000s,10], s2d[0.001s,1], nnc[0.001s,1800], ad[0.004s,36], nni[0.008s,1800], H2D[0.024s,9], nn
                                                                                                      ba[0.040s,36], nncf[0.050s,1800], az[0.128s,1800], sdrixn[0.246s,1800]
  5  gpu_maxpooling_backward [100:2-101:70]                0.428     36                      nnc[0.000s,36], nni[0.000s,36], nnmf[0.001s,36], nnmb[0.001s,36], ad[0.003s,36]
  6  sp_rblk [cnn.dml 15:0-15:11]                          0.278      1
  7  gpu_maxpooling [63:2-64:62]                           0.231     36                                       nnc[0.000s,36], nni[0.000s,36], nnmf[0.001s,36], ad[0.002s,36]
  8  rand [./nn/layers/conv2d_builtin.dml 158:6-158:45]    0.081      4                                                                      rlswr[0.001s,0], aqmd[0.001s,4]
  9  gpu_uark+ [42:2-42:56]                                0.069     36                     aqrs[0.000s,36], H2D[0.002s,36], ad[0.003s,36], rrowk[0.003s,36], s2d[0.024s,36]
…