@nakul02 2017-09-20
Thank you for throwing extensive benchmarking numbers. 
Everything LGTM 👍 

When you get the chance, please run all the GPU tests, either on your own machine or Jenkins.